+++ R/geom_stepribbon.R
@@ -1,0 +1,40 @@
+# Taken from the \code{RmcdrPlugin.KMggplot2}, slightly modified
+geom_stepribbon <- function(
+  mapping     = NULL,
+  data        = NULL,
+  stat        = "identity",
+  position    = "identity",
+  na.rm       = FALSE,
+  show.legend = NA,
+  inherit.aes = TRUE, ...) {
+
+  ggplot2::layer(
+    data        = data,
+    mapping     = mapping,
+    stat        = stat,
+    geom        = GeomStepribbon,
+    position    = position,
+    show.legend = show.legend,
+    inherit.aes = inherit.aes,
+    params      = list(na.rm = na.rm, ... )
+  )
+
+}
+
+GeomStepribbon <- ggplot2::ggproto(
+  "GeomStepribbon", ggplot2::GeomRibbon,
+
+  extra_params = c("na.rm"),
+
+  draw_group = function(data, panel_scales, coord, na.rm = FALSE) {
+
+    if (na.rm) data <- data[complete.cases(data[c("x", "ymin", "ymax")]), ]
+    data   <- rbind(data, data)
+    data   <- data[order(data$x), ]
+    data$x <- c(data$x[2:nrow(data)], NA)
+    data   <- data[complete.cases(data["x"]), ]
+    ggplot2::GeomRibbon$draw_group(data, panel_scales, coord, na.rm = FALSE)
+
+  }
+
+)